# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.ResolveAlertRequest:
    properties:
      notes:
        example: False positive - legitimate admin access
        type: string
    type: object
  handlers.SecurityDashboardResponse:
    properties:
      ml_dashboard:
        additionalProperties: true
        type: object
      recent_alerts:
        items:
          $ref: '#/definitions/models.AnomalyAlert'
        type: array
      security_metrics:
        items:
          $ref: '#/definitions/models.SecurityMetrics'
        type: array
      summary:
        additionalProperties: true
        type: object
    type: object
  handlers.StandardErrorResponse:
    properties:
      error:
        example: Invalid email or password
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  models.AnomalyAlert:
    properties:
      anomaly_reasons:
        description: JSON array of reasons
        type: string
      anomaly_score:
        type: number
      created_at:
        type: string
      id:
        type: string
      ip_address:
        type: string
      is_resolved:
        type: boolean
      notes:
        type: string
      request_log:
        allOf:
        - $ref: '#/definitions/models.RequestLog'
        description: Relationships
      request_log_id:
        type: string
      resolved_at:
        type: string
      resolved_by:
        type: string
      resolved_by_user:
        $ref: '#/definitions/models.User'
      risk_level:
        description: '''low'', ''medium'', ''high'', ''critical'''
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
  models.CartItem:
    properties:
      cart:
        allOf:
        - $ref: '#/definitions/models.ShoppingCart'
        description: Relationships
      cart_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      quantity:
        type: integer
      updated_at:
        type: string
    type: object
  models.Comment:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      rating:
        description: 1-5 star rating
        type: integer
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: string
    type: object
  models.Discount:
    properties:
      category:
        description: Nullable for product-specific discounts
        type: string
      created_at:
        type: string
      discount_type:
        description: '''percentage'', ''fixed_amount'''
        type: string
      discount_value:
        description: Percentage (0-100) or fixed amount
        type: number
      end_date:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      max_discount_amount:
        description: Maximum discount amount (for percentage)
        type: number
      min_order_amount:
        description: Minimum order amount to apply discount
        type: number
      product:
        allOf:
        - $ref: '#/definitions/models.Product'
        description: Relationships
      product_id:
        description: Nullable for category-wide discounts
        type: string
      start_date:
        type: string
      updated_at:
        type: string
      usage_count:
        type: integer
      usage_limit:
        description: 0 = unlimited
        type: integer
    type: object
  models.Favorite:
    properties:
      created_at:
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: string
    type: object
  models.Order:
    properties:
      created_at:
        type: string
      id:
        type: string
      order_items:
        items:
          $ref: '#/definitions/models.OrderItem'
        type: array
      status:
        type: string
      total:
        type: number
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: string
    type: object
  models.OrderItem:
    properties:
      created_at:
        type: string
      id:
        type: string
      order:
        allOf:
        - $ref: '#/definitions/models.Order'
        description: Relationships
      order_id:
        type: string
      price:
        type: number
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      quantity:
        type: integer
    type: object
  models.Product:
    properties:
      cart_items:
        items:
          $ref: '#/definitions/models.CartItem'
        type: array
      category:
        type: string
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      created_at:
        type: string
      description:
        type: string
      discounts:
        items:
          $ref: '#/definitions/models.Discount'
        type: array
      favorites:
        items:
          $ref: '#/definitions/models.Favorite'
        type: array
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
      order_items:
        description: Relationships
        items:
          $ref: '#/definitions/models.OrderItem'
        type: array
      price:
        type: number
      product_views:
        items:
          $ref: '#/definitions/models.ProductView'
        type: array
      recommendations:
        items:
          $ref: '#/definitions/models.Recommendation'
        type: array
      stock:
        type: integer
      tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      updated_at:
        type: string
      upvotes:
        items:
          $ref: '#/definitions/models.Upvote'
        type: array
      user_interactions:
        items:
          $ref: '#/definitions/models.UserInteraction'
        type: array
    type: object
  models.ProductView:
    properties:
      created_at:
        description: Changed from Timestamp to CreatedAt for consistency
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      session_id:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: string
    type: object
  models.Recommendation:
    properties:
      algorithm_type:
        description: '''collaborative'', ''content_based'', ''hybrid'''
        type: string
      created_at:
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      score:
        type: number
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: string
    type: object
  models.RequestLog:
    properties:
      id:
        type: string
      ip_address:
        type: string
      method:
        type: string
      path:
        type: string
      query_params:
        type: string
      request_size:
        type: integer
      response_size:
        type: integer
      response_time:
        description: in milliseconds
        type: number
      session_id:
        type: string
      status_code:
        type: integer
      timestamp:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_agent:
        type: string
      user_id:
        type: string
    type: object
...
  models.SecurityMetrics:
    properties:
      anomalous_requests:
        type: integer
      avg_response_time:
        type: number
      blocked_requests:
        type: integer
      created_at:
        type: string
      date:
        type: string
      error_rate:
        type: number
      high_risk_requests:
        type: integer
      id:
        type: string
      total_requests:
        type: integer
      unique_ips:
        type: integer
      updated_at:
        type: string
    type: object
  models.ShoppingCart:
    properties:
      cart_items:
        items:
          $ref: '#/definitions/models.CartItem'
        type: array
      created_at:
        type: string
      id:
        type: string
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: string
    type: object
  models.Tag:
    properties:
      color:
        description: Hex color code
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      products:
        description: Relationships
        items:
          $ref: '#/definitions/models.Product'
        type: array
    type: object
  models.Upvote:
    properties:
      created_at:
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: string
    type: object
  models.User:
    properties:
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      created_at:
        type: string
      email:
        type: string
      favorites:
        items:
          $ref: '#/definitions/models.Favorite'
        type: array
      id:
        type: string
      name:
        type: string
      orders:
        description: Relationships
        items:
          $ref: '#/definitions/models.Order'
        type: array
      recommendations:
        items:
          $ref: '#/definitions/models.Recommendation'
        type: array
      shopping_cart:
        $ref: '#/definitions/models.ShoppingCart'
      updated_at:
        type: string
      upvotes:
        items:
          $ref: '#/definitions/models.Upvote'
        type: array
      user_interactions:
        items:
          $ref: '#/definitions/models.UserInteraction'
        type: array
    type: object
  models.UserInteraction:
    properties:
      created_at:
        description: Changed from Timestamp to CreatedAt for consistency
        type: string
      id:
        type: string
      interaction_type:
        description: '''view'', ''cart_add'', ''purchase'', ''like'''
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      session_id:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: string
    type: object
host: localhost:8081
info:
  contact:
    email: support@bachelor-ecommerce.com
    name: API Support
    url: http://www.swagger.io/support
  description: ML-Powered E-commerce Platform API with advanced recommendation system,
    intelligent search, and comprehensive analytics
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Bachelor E-commerce API
  version: "1.0"
paths:
...
  /security/alerts:
    get:
      consumes:
      - application/json
      description: Get recent security alerts with filtering options
      parameters:
      - default: 50
        description: Number of alerts to return
        in: query
        maximum: 1000
        minimum: 1
        name: limit
        type: integer
      - description: Filter by risk level
        enum:
        - low
        - medium
        - high
        - critical
        in: query
        name: risk_level
        type: string
      - description: Filter by resolution status
        in: query
        name: resolved
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Security alerts data
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Security Alerts
      tags:
      - Security
  /security/alerts/{alert_id}/resolve:
    post:
      consumes:
      - application/json
      description: Mark a security alert as resolved with optional notes
      parameters:
      - description: Alert ID
        format: uuid
        in: path
        name: alert_id
        required: true
        type: string
      - description: Resolution details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ResolveAlertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Alert resolved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - invalid alert ID or request body
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "404":
          description: Alert not found
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
      security:
      - BearerAuth: []
      summary: Resolve Security Alert
      tags:
      - Security
  /security/dashboard:
    get:
      consumes:
      - application/json
      description: Get comprehensive security dashboard with anomaly detection data
        and real-time threat monitoring
      produces:
      - application/json
      responses:
        "200":
          description: Security dashboard data
          schema:
            $ref: '#/definitions/handlers.SecurityDashboardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Security Dashboard
      tags:
      - Security
  /security/insights:
    get:
      consumes:
      - application/json
      description: Get detailed insights and analysis about detected anomalies and
        attack patterns
      produces:
      - application/json
      responses:
        "200":
          description: Anomaly insights data
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Anomaly Insights
      tags:
      - Security
...
  /security/patterns:
    get:
      consumes:
      - application/json
      description: Get information about detectable attack patterns and security threats
      produces:
      - application/json
      responses:
        "200":
          description: Attack patterns information
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Attack Patterns
      tags:
      - Security
  /security/simulate:
    get:
      consumes:
      - application/json
      description: Simulate different types of attacks for demonstration and testing
        purposes
      parameters:
      - description: Attack type
        enum:
        - sql_injection
        - xss
        - brute_force
        - ddos
        example: sql_injection
        in: query
        name: attack_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Attack simulation results
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - invalid attack type
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
      security:
      - BearerAuth: []
      summary: Simulate Attack
      tags:
      - Security
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...